pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "think-tank"
version = "0.1.0"
description = "LLM-powered multi-expert structured debate tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Eugene Nayshtetik"},
]
dependencies = [
    "anthropic>=0.39.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
think-tank = "think_tank.cli:main"

[tool.setuptools.packages.find]
include = ["think_tank*"]
